New feature: ChooseOrCreateNew
Add form component that provides dropdown field with passable array, and
input for creating new value.

This component has been created to allow the user to select a pre-existing
entity in the app i.e. agent to create a record against, or they can
provide a string which will create a new entity on the server, against which
the item they're creating will then be associated with.

This functionality needs to be abstracted on the server side as there will
be several situations in which it will need to be used.